move some methods and field
